+++ openarm_hardware/src/v10_simple_hardware.cpp
@@ -13,6 +13,7 @@
 // limitations under the License.
 
 #include "openarm_hardware/v10_simple_hardware.hpp"
+#include "openarm_hardware/config_yaml.hpp"
 
 #include <yaml-cpp/yaml.h>
 
@@ -51,154 +52,6 @@ std::string error_code_to_string(uint8_t error_code) {
   }
 }
 }  // namespace
-
-// YAML conversion implementations
-namespace YAML {
-
-openarm::damiao_motor::MotorType parse_motor_type(const std::string& type_str) {
-  if (type_str == "DM4310") return openarm::damiao_motor::MotorType::DM4310;
-  if (type_str == "DM4340") return openarm::damiao_motor::MotorType::DM4340;
-  if (type_str == "DM8009") return openarm::damiao_motor::MotorType::DM8009;
-  if (type_str == "DMH6215") return openarm::damiao_motor::MotorType::DMH6215;
-  if (type_str == "DM8006") return openarm::damiao_motor::MotorType::DM8006;
-  if (type_str == "DM6006") return openarm::damiao_motor::MotorType::DM6006;
-
-  throw std::runtime_error("Unknown motor type: " + type_str);
-}
-
-bool convert<openarm_hardware::MotorConfig>::decode(
-    const Node& node, openarm_hardware::MotorConfig& config) {
-  if (!node.IsMap()) {
-    return false;
-  }
-
-  try {
-    config.name = node["name"].as<std::string>();
-
-    // Parse motor type
-    config.type = parse_motor_type(node["type"].as<std::string>());
-    config.send_can_id = node["send_can_id"].as<uint32_t>();
-    config.recv_can_id = node["recv_can_id"].as<uint32_t>();
-    config.kp = node["kp"].as<double>();
-    config.kd = node["kd"].as<double>();
-
-    return true;
-  } catch (const std::exception& e) {
-    RCLCPP_ERROR(rclcpp::get_logger("OpenArm_v10HW"),
-                 "Failed to parse motor config: %s", e.what());
-    return false;
-  }
-}
-
-bool convert<openarm_hardware::GripperConfig>::decode(
-    const Node& node, openarm_hardware::GripperConfig& config) {
-  if (!node.IsMap()) {
-    return false;
-  }
-
-  try {
-    config.name = node["name"].as<std::string>();
-
-    config.motor_type = parse_motor_type(node["motor_type"].as<std::string>());
-
-    config.send_can_id = node["send_can_id"].as<uint32_t>();
-    config.recv_can_id = node["recv_can_id"].as<uint32_t>();
-    config.kp = node["kp"].as<double>();
-    config.kd = node["kd"].as<double>();
-
-    // Position mapping
-    config.closed_position = node["closed_position"].as<double>();
-    config.open_position = node["open_position"].as<double>();
-    config.motor_closed_radians = node["motor_closed_radians"].as<double>();
-    config.motor_open_radians = node["motor_open_radians"].as<double>();
-
-    return true;
-  } catch (const std::exception& e) {
-    RCLCPP_ERROR(rclcpp::get_logger("OpenArm_v10HW"),
-                 "Failed to parse gripper config: %s", e.what());
-    return false;
-  }
-}
-
-bool convert<openarm_hardware::ControllerConfig>::decode(
-    const Node& node, openarm_hardware::ControllerConfig& config) {
-  if (!node.IsMap()) {
-    return false;
-  }
-
-  try {
-    auto logger = rclcpp::get_logger("OpenArm_v10HW");
-
-    // Load CAN interface configuration (required)
-    if (!node["can_iface"]) {
-      throw std::runtime_error("Missing required field: can_iface");
-    }
-    config.can_iface = node["can_iface"].as<std::string>();
-
-    if (!node["can_fd"]) {
-      throw std::runtime_error("Missing required field: can_fd");
-    }
-    config.can_fd = node["can_fd"].as<bool>();
-
-    // Load arm configuration (required)
-    if (!node["arm"]) {
-      throw std::runtime_error("Missing required field: arm");
-    }
-    const YAML::Node& arm_node = node["arm"];
-
-    if (!arm_node.IsSequence()) {
-      RCLCPP_ERROR(logger, "arm configuration must be a sequence");
-      return false;
-    }
-
-    for (const auto& motor_node : arm_node) {
-      openarm_hardware::MotorConfig motor_config =
-          motor_node.as<openarm_hardware::MotorConfig>();
-      config.arm_joints.push_back(motor_config);
-    }
-
-    RCLCPP_INFO(logger, "Loaded %zu arm motor configurations from YAML",
-                config.arm_joints.size());
-
-    // Load auxiliary motors configuration (also a list)
-    if (node["auxiliary"]) {
-      const YAML::Node& aux_node = node["auxiliary"];
-
-      if (!aux_node.IsSequence()) {
-        RCLCPP_INFO(logger, "auxiliary configuration must be a sequence");
-        return false;
-      }
-
-      for (const auto& motor_node : aux_node) {
-        openarm_hardware::MotorConfig motor_config =
-            motor_node.as<openarm_hardware::MotorConfig>();
-        config.arm_joints.push_back(motor_config);
-      }
-
-      RCLCPP_INFO(logger, "Loaded %zu auxiliary motor configurations",
-                  aux_node.size());
-    }
-
-    // Load gripper configuration
-    if (node["gripper"]) {
-      config.gripper_joint =
-          node["gripper"].as<openarm_hardware::GripperConfig>();
-      RCLCPP_INFO(logger, "Loaded gripper configuration");
-    }
-
-    return true;
-  } catch (const YAML::Exception& e) {
-    RCLCPP_ERROR(rclcpp::get_logger("OpenArm_v10HW"),
-                 "Failed to parse YAML configuration: %s", e.what());
-    return false;
-  } catch (const std::exception& e) {
-    RCLCPP_ERROR(rclcpp::get_logger("OpenArm_v10HW"),
-                 "Failed to load YAML configuration: %s", e.what());
-    return false;
-  }
-}
-
-}  // namespace YAML
 
 namespace openarm_hardware {
